Use pg_dump --quote-all-identifiers option
In 5893ffa79c03824f34ae3d37f211381fd1c17283, "over" was reclassified
as an unreserved keyword, which means it can now be used without
quotes as a table or column name (e.g., `CREATE TABLE over(over
int)`). However, when upgrading databases by piping a pg_dump into
a pg_restore, this can cause problems if upgrading a database from
before this change (which requires 'over' to be quoted when used as
a column name) to one with this change (where pg_dump does not think
that 'over' has to be quoted anymore).

We can't just use an old source database's pg_dump when upgrading,
because the newer target database may have introduced *other* reserved
keywords that the old pg_dump will not know to quote.

So we just quote the heck out of everything. The only downside is more
quotes in the text of the dump, which seems manageable.